requirements.txt
Django==4.2.1
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0 